AOC-2022 17-12/1 solution to riddle 17 part 1

+++ src/main/java/de/barryallenofearth/adventofcode2022/riddle/q/rocks/common/model/Coordinates.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class Coordinates {
-	private int x;
+	private long x;
 
-	private int y;
+	private long y;
 }
